Reorganized files and updated server for GET and POST requests and updated client for POST request. #2
