+++ server/index.js
@@ -19,20 +19,32 @@ const db = new pg.Pool({
 });
 
 app.post('/api/memories', (req, res, next) => {
-  const { placeVisited, date, favoriteMoments } = req.body;
+  const { placeVisited, date, favoriteMoments, lat, long } = req.body;
   if (!placeVisited || !date) {
     throw new ClientError(400, 'location and date are required fields');
   }
   const sql = `
-    insert into "memories" ("placeVisited", "date", "favoriteMoments")
-    values ($1, $2, $3)
+    insert into "memories" ("placeVisited", "date", "favoriteMoments", "lat", "long")
+    values ($1, $2, $3, $4, $5)
     returning *
   `;
-  const params = [placeVisited, date, favoriteMoments];
+  const params = [placeVisited, date, favoriteMoments, lat, long];
   db.query(sql, params)
     .then(result => {
       const [memory] = result.rows;
       res.status(201).json(memory);
+    })
+    .catch(err => next(err));
+});
+
+app.get('/api/memories', (req, res, next) => {
+  const sql = `
+    select *
+      from "memories"
+  `;
+  db.query(sql)
+    .then(result => {
+      res.json(result.rows);
     })
     .catch(err => next(err));
 });
